api/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
mysql-connector-python==8.0.32
pydantic==2.3.0
sqlparse==0.4.4